+++ Assets/Code/AbilitySystem/Abilities/Bombard.cs
@@ -22,7 +22,7 @@ namespace Assets.Code.AbilitySystem.Abilities
         private float _projectilesCount;
 
         public Bombard(AbilityConfig config, Dictionary<AbilityType, int> abilityUnlockLevel, Transform launchPoint,
-            Dictionary<AbilityType, float> damageDealt, Dictionary<AbilityType, int> killCount, int level = 1) : base(config, abilityUnlockLevel, level)
+            Dictionary<AbilityType, int> damageDealt, Dictionary<AbilityType, int> killCount, int level = 1) : base(config, abilityUnlockLevel, level)
         {
             AbilityStats stats = config.ThrowIfNull().GetStats(level);
             _launchPoint = launchPoint.ThrowIfNull();
