+++ Assets/Code/AbilitySystem/Abilities/IceStaff.cs
@@ -23,7 +23,7 @@ namespace Assets.Code.AbilitySystem.Abilities
         private int _projectilesCount;
 
         public IceStaff(AbilityConfig config, Dictionary<AbilityType, int> abilityUnlockLevel, Transform heroCenterPoint,
-            Dictionary<AbilityType, float> damageDealt, Dictionary<AbilityType, int> killCount, int level = 1) : base(config, abilityUnlockLevel, level)
+            Dictionary<AbilityType, int> damageDealt, Dictionary<AbilityType, int> killCount, int level = 1) : base(config, abilityUnlockLevel, level)
         {
             AbilityStats stats = config.ThrowIfNull().GetStats(level);
             _heroCenter = heroCenterPoint.ThrowIfNull();
